Moved resource handling to storage
