+++ plugins/org.carat.plugin/src/org/carat20/client/Carat.java
@@ -1,17 +1,11 @@
 package org.carat20.client;
 
 import android.content.Context;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.drawable.BitmapDrawable;
-import android.graphics.drawable.Drawable;
-import android.util.Base64;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CallbackContext;
 import org.json.JSONArray;
 import org.json.JSONException;
 import android.util.Log;
-import java.io.ByteArrayOutputStream;
 import java.util.Arrays;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaWebView;
@@ -205,11 +199,10 @@ public class Carat extends CordovaPlugin {
         Log.v("Converting hog/bug reports to JSON", Arrays.toString(reports));
         JSONArray results = new JSONArray();
         for(SimpleHogBug s : reports){
-                String packageName = s.getAppName();
-                JSONObject reportObject = new JSONObject()
+                JSONObject app = new JSONObject()
                         .put("label", s.getAppLabel())
-                        .put("icon", this.getBase64Icon(packageName))
-                        .put("name", packageName)
+                        .put("icon", s.getAppIcon())
+                        .put("name", s.getAppName())
                         .put("priority",s.getAppPriority())
                         .put("benefit",s.getBenefitText())
                         .put("samples", s.getSamples())
@@ -217,7 +210,7 @@ public class Carat extends CordovaPlugin {
                         .put("expected", s.getExpectedValue())
                         .put("expectedWithout", s.getExpectedValueWithout())
                         .put("type", s.getType());
-                results.put(reportObject);
+                results.put(app);
         }
         return results;
     }
@@ -241,39 +234,5 @@ public class Carat extends CordovaPlugin {
                 .put("similarApps", r.similarApps)
                 .put("similarAppsWithout", r.similarAppsWithout);
         return results;
-    }
-    
-    // Drawable resource encoding
-    
-    /**
-     * Return base64 encoded icon PNG from application package.
-     * @param packageName Package name.
-     * @return Base 64 encoded PNG or an empty string.
-     */
-    private String getBase64Icon(String packageName){
-        try{
-            Drawable d = context.getPackageManager().getApplicationIcon(packageName);
-            return "data:image/png;base64,"+ encodeIcon(d);
-        } catch (PackageManager.NameNotFoundException e){
-            return "";
-        }
-    }
-    /**
-     * Converts a drawable resource to base64 encoded PNG image.
-     * @param icon Drawable image resource.
-     * @return Base64 representation of PNG compressed bitmap.
-     */
-    public static String encodeIcon(Drawable icon){
-            if(icon == null) return "";
-            
-            BitmapDrawable bmDrawable = ((BitmapDrawable) icon);
-            Bitmap bitmap = bmDrawable.getBitmap();
-            
-            bitmap = Bitmap.createScaledBitmap(bitmap, 48, 48, true);
-            ByteArrayOutputStream outStream = new ByteArrayOutputStream();
-            bitmap.compress(Bitmap.CompressFormat.PNG, 100, outStream);
-            byte[] bitmapByte = outStream.toByteArray();
-
-            return Base64.encodeToString(bitmapByte,Base64.DEFAULT);
     }
 }
